requirements.txt
bs4~=0.0.1
lxml
sendgrid~=6.4.4
beautifulsoup4~=4.9.1
selenium~=3.141.0
webdriver-manager

# for GDrive usage
gspread~=3.6.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
